requirements.txt
streamlit
google-cloud-bigquery
pandas
db-dtypes
google-auth
google-auth-oauthlib
google-auth-httplib2
google-genai
langchain
faiss-cpu
sentence-transformers
langchain-community
langchain-text-splitters